+++ src/main/java/ru/vyarus/guice/persist/orient/db/transaction/TxConfig.java
@@ -1,7 +1,9 @@
 package ru.vyarus.guice.persist.orient.db.transaction;
 
 import com.google.common.base.MoreObjects;
+import com.orientechnologies.orient.core.db.ODatabaseRecordThreadLocal;
 import com.orientechnologies.orient.core.tx.OTransaction;
+import com.orientechnologies.orient.core.tx.OTransactionNoTx;
 
 import java.util.Collections;
 import java.util.List;
@@ -13,10 +15,13 @@ import java.util.List;
  * @author Vyacheslav Rusakov
  * @since 25.07.2014
  */
+@SuppressWarnings("PMD.AvoidFieldNameMatchingMethodName")
 public class TxConfig {
     private List<Class<? extends Exception>> rollbackOn;
     private List<Class<? extends Exception>> ignore;
     private OTransaction.TXTYPE txtype;
+    // transaction is managed externally
+    private boolean external;
 
     public TxConfig() {
         this(null, null, null);
@@ -76,10 +81,59 @@ public class TxConfig {
         return txtype;
     }
 
+    /**
+     * External transactions allow re-using already existing and bound to thread database instance.
+     * But in this case no commit or rollback will be performed.
+     *
+     * @return true when transaction is managed externally, false otherwise
+     */
+    public boolean isExternal() {
+        return external;
+    }
+
     @Override
     public String toString() {
-        return "{type " + txtype
+        return external ? "{external}" : "{type " + txtype
                 + (rollbackOn.isEmpty() ? "" : " rollbackOn " + rollbackOn)
                 + (ignore.isEmpty() ? "" : " ignore " + ignore) + "}";
+    }
+
+    /**
+     * External transaction is re-use of database instance already bound to thread. For example,
+     * <pre>{@code
+     *      // transaction opened manually
+     *      ODatabaseDocumentTx db = new ODatabaseDocumentTx();
+     *      // only after that it could be detected
+     *      db.activateOnCurrentThread();
+     *
+     *      // here we can use external transaction
+     *
+     *      // connection closed manually
+     *      db.close();
+     * }</pre>
+     * <p>
+     * External transaction is a way to re-use already existing connection (e.g. managed by orient itself) in
+     * guice beans. The case must be applied with care: as connection is managed outside of guice, external
+     * transaction scope must be exactly between manual db open and close (otherwise you will have error due
+     * to not bound db instance). Rollback and commit actions will not affect connection.
+     * <p>
+     * Type of transaction is resolved from external connection.
+     * <p>
+     * One example usage case is non blocking async queries listener: there orient create connection in separate
+     * thread and calls listener multiple times, so if usual transaction would be used inside listener,
+     * it will override existing orient transaction and orient logic execution will break.
+     *
+     * @return external transaction config
+     */
+    public static TxConfig external() {
+        final TxConfig cfg = new TxConfig();
+        cfg.external = true;
+        // currently only only notx and optimistic transactions supported
+        cfg.txtype = ODatabaseRecordThreadLocal.INSTANCE.get()
+                .getTransaction() instanceof OTransactionNoTx
+                ? OTransaction.TXTYPE.NOTX : OTransaction.TXTYPE.OPTIMISTIC;
+        cfg.rollbackOn.clear();
+        cfg.ignore.clear();
+        return cfg;
     }
 }