+++ src/main/java/ru/vyarus/guice/persist/orient/db/transaction/TxType.java
@@ -5,16 +5,17 @@ import com.orientechnologies.orient.core.tx.OTransaction;
 import java.lang.annotation.*;
 
 /**
- * <p>Defines orient transaction type to use within current transaction.
+ * Defines orient transaction type to use within current transaction.
  * Must be used together with guice @Transactional annotation.
  * May be defined on type or method, but not necessary near @Transactional annotation. E.g. @Transactional could be
  * set on type and @TxType on single bean method, which will change transaction type only for this method.
- * </p>
- * To switch of transaction within unit of work (defined by @Transactional annotation) use
- * {@code OTransaction.TXTYPE.NOTX}
- * <p>Additional annotation was chosen in order to not introduce new annotation for transaction and re-use
+ * <p>
+ * To switch off transaction within unit of work (defined by @Transactional annotation) use
+ * {@link OTransaction.TXTYPE#NOTX}
+ * <p>
+ * Additional annotation was chosen in order to not introduce new annotation for transaction and re-use
  * guice-persist one. Moreover, transaction type definition should be a rear case (because if its not you should
- * change default transaction type in module and use @TxType annotation just for rear cases</p>
+ * change default transaction type in module and use @TxType annotation just for rear cases.
  *
  * @author Vyacheslav Rusakov
  * @since 25.07.2014
@@ -25,5 +26,9 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
 public @interface TxType {
+
+    /**
+     * @return type of transaction
+     */
     OTransaction.TXTYPE value();
 }