+++ CHANGELOG.md
@@ -13,6 +13,9 @@
     - required orient OLiveResultListener parameter must be used with @Listen annotation
     - custom LiveQueryListener interface can be used instead of OLiveResultListener to apply automatic result conversion 
       (with RecordConverter to mimic the same behaviour as for usual method result).    
+* Support external connection (thread bound) re-use: when transaction started with TxConfig.external() thread bound
+  connection used instead of new connection. Commits and rollbacks are not applied automatically: supposed that manual 
+  connection is completely managed externally. Useful in case when already existing connection must be (re)used in guice.
 
 ### 3.2.0 (2016-09-25)
 * Update to orient 2.1